Switch to the master branch for urdfdom
In https://github.com/ros/urdfdom/pull/158 and https://github.com/ros/urdfdom/pull/161,
we merged the ros2 branch onto the master branch so there would be single,
unified release branch for all of urdfdom.  Now that that is done, we should switch
to using the master branch of urdfdom.