Bump versions of ROS2 dependencies (#135)
This commit bumps the version numbers of dependencies in Space ROS that are
also dependencies of ROS2, to match those in the latest patch to ROS2's Humble
release (release-humble-20231122).

The package ament_lint is kept pointing to the spaceros branch, since we don't
have a way to bring it into sync with the version of ament_lint pinned by ROS2
before our next release.