Move ubuntu version to noble/24.04, used by ROS 2 Jazzy Jalisco (#157)
We want to move to Jazzy in Space ROS, as it includes improvements that will be
beneficial to the Space ROS project.

ROS 2 Jazzy is, by default, built on top of Ubuntu Noble (24.04).

This commit changes the Ubuntu version used in Space ROS to match that of
standard ROS 2 Jazzy.